style) utils 파일 내 함수 관심사별 구분 작업 진행
